+++ src/COPIERTemplate.jl
@@ -10,6 +10,8 @@ module COPIERTemplate
 
 include("Copier.jl")
 
+using TOML: TOML
+
 """
     generate(dst_path[, data]; kwargs...)
     generate(src_path, dst_path[, data]; true, kwargs...)
@@ -24,11 +26,50 @@ The `data` argument is a dictionary of answers (values) to questions (keys) that
 The keyword arguments are passed directly to the internal [`Copier.copy`](@ref).
 """
 function generate(src_path, dst_path, data::Dict = Dict(); kwargs...)
+  # If there are answers in the destionation path, skip guessing the answers
+  if !isfile(joinpath(dst_path, ".copier-answers")) && isdir(dst_path)
+    existing_data = _read_data_from_existing_path(dst_path)
+    for (key, value) in existing_data
+      @info "Inferred $key=$value from destination path"
+      if haskey(data, key)
+        @info "  Being overriden by supplied $key=$(data[key]) value"
+      end
+    end
+    data = merge(existing_data, data)
+  end
   Copier.copy(src_path, dst_path, data; kwargs...)
 end
 
 function generate(dst_path, data::Dict = Dict(); kwargs...)
   generate("https://github.com/abelsiqueira/COPIERTemplate.jl", dst_path, data; kwargs...)
+end
+
+"""
+    data = _read_data_from_existing_path(dst_path)
+
+Reads the destination folder to figure out some answers.
+"""
+function _read_data_from_existing_path(dst_path)
+  data = Dict{String, Any}()
+  if isfile(joinpath(dst_path, "Project.toml"))
+    toml_data = TOML.parsefile(joinpath(dst_path, "Project.toml"))
+    for (toml_key, copier_key) in [("name", "PackageName"), ("uuid", "PackageUUID")]
+      if haskey(toml_data, toml_key)
+        data[copier_key] = toml_data[toml_key]
+      end
+    end
+    # Author capture is limited and does not handle multiple authors. See #118 for more information.
+    if haskey(toml_data, "authors")
+      author_regex = r"^(.*) <(.*)>(?: and contributors)?"
+      m = match(author_regex, toml_data["authors"][1])
+      if !isnothing(m)
+        data["AuthorName"] = m[1]
+        data["AuthorEmail"] = m[2]
+      end
+    end
+  end
+
+  return data
 end
 
 end