+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning].
 - New question: SimplifiedPRTest to simplify the testing on Pull Requests (#105)
 - New question: AddAllcontributors to add a section and config for <https://allcontributors.org> (#26)
 - `copy`, `recopy` and `update` from the copier API (#142)
+- When applying to existing projects, read Project.toml to infer a few values (#116)
 
 ### Changed